+++ circonus/metric.go
@@ -37,21 +37,8 @@ func (m *circonusMetric) ParseConfig(id string, d *schema.ResourceData) error {
 		m.Status = metricActiveToAPIStatus(v.(bool))
 	}
 
-	if v, found := d.GetOk(metricTagsAttr); found {
-		m.Tags = derefStringList(flattenSet(v.(*schema.Set)))
-	}
-
 	if v, found := d.GetOk(metricTypeAttr); found {
 		m.Type = v.(string)
-	}
-
-	if v, found := d.GetOk(metricUnitAttr); found {
-		s := v.(string)
-		m.Units = &s
-	}
-
-	if m.Units != nil && *m.Units == "" {
-		m.Units = nil
 	}
 
 	return nil
@@ -68,21 +55,8 @@ func (m *circonusMetric) ParseConfigMap(id string, attrMap map[string]interface{
 		m.Status = metricActiveToAPIStatus(v.(bool))
 	}
 
-	if v, found := attrMap[metricTagsAttr]; found {
-		m.Tags = derefStringList(flattenSet(v.(*schema.Set)))
-	}
-
 	if v, found := attrMap[metricTypeAttr]; found {
 		m.Type = v.(string)
-	}
-
-	if v, found := attrMap[metricUnitAttr]; found {
-		s := v.(string)
-		m.Units = &s
-	}
-
-	if m.Units != nil && *m.Units == "" {
-		m.Units = nil
 	}
 
 	return nil
@@ -93,9 +67,7 @@ func (m *circonusMetric) SaveState(d *schema.ResourceData) error {
 
 	_ = d.Set(metricActiveAttr, metricAPIStatusToBool(m.Status))
 	_ = d.Set(metricNameAttr, m.Name)
-	_ = d.Set(metricTagsAttr, tagsToState(apiToTags(m.Tags)))
 	_ = d.Set(metricTypeAttr, m.Type)
-	_ = d.Set(metricUnitAttr, indirect(m.Units))
 
 	return nil
 }
@@ -150,31 +122,8 @@ func metricChecksum(m interfaceMap) int {
 		fmt.Fprint(b, v.(string))
 	}
 
-	if v, found := m[metricTagsAttr]; found {
-		tags := derefStringList(flattenSet(v.(*schema.Set)))
-		for _, tag := range tags {
-			fmt.Fprint(b, tag)
-		}
-	}
-
 	if v, found := m[metricTypeAttr]; found {
 		fmt.Fprint(b, v.(string))
-	}
-
-	if v, found := m[metricUnitAttr]; found {
-		if v != nil {
-			var s string
-			switch v := v.(type) {
-			case string:
-				s = v
-			case *string:
-				s = *v
-			}
-
-			if s != "" {
-				fmt.Fprint(b, s)
-			}
-		}
 	}
 
 	s := b.String()