+++ circonus/metric_test.go
@@ -1,19 +1,18 @@
 package circonus
 
-import "testing"
+import (
+	"testing"
+)
 
 func Test_MetricChecksum(t *testing.T) {
-	unit := "qty"
 	m := interfaceMap{
 		string(metricActiveAttr): true,
 		string(metricNameAttr):   "asdf",
-		string(metricTagsAttr):   tagsToState(apiToTags([]string{"foo", "bar"})),
 		string(metricTypeAttr):   "json",
-		string(metricUnitAttr):   &unit,
 	}
 
 	csum := metricChecksum(m)
-	if csum != 4250221491 {
+	if csum != 4074128010 {
 		t.Fatalf("Checksum mismatch")
 	}
 }