+++ circonus/provider.go
@@ -99,7 +99,6 @@ func Provider() terraform.ResourceProvider {
 			"circonus_dashboard":      resourceDashboard(),
 			"circonus_maintenance":    resourceMaintenance(),
 			"circonus_metric":         resourceMetric(),
-			"circonus_metric_cluster": resourceMetricCluster(),
 			"circonus_rule_set":       resourceRuleSet(),
 			"circonus_rule_set_group": resourceRuleSetGroup(),
 			"circonus_worksheet":      resourceWorksheet(),