+++ circonus/resource_circonus_check_cloudwatch_test.go
@@ -46,159 +46,73 @@ func TestAccCirconusCheckCloudWatch_basic(t *testing.T) {
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "period", "60s"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.#", "17"),
 
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.name", "ReadLatency"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.unit", "seconds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.name", "TransactionLogsGeneration"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.unit", ""),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.name", "WriteIOPS"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.unit", "iops"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.name", "FreeStorageSpace"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.unit", ""),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.name", "WriteLatency"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.unit", "seconds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.name", "DatabaseConnections"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.unit", "connections"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.name", "FreeableMemory"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.unit", "bytes"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.name", "MaximumUsedTransactionIDs"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.unit", ""),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.name", "ReadThroughput"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.unit", "bytes"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.name", "ReadIOPS"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.unit", "iops"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.name", "NetworkReceiveThroughput"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.unit", "bytes"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.name", "TransactionLogsDiskUsage"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.unit", "bytes"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.active", "true"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.name", "CPUUtilization"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.tags.4259413593", "source:cloudwatch"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.0.unit", "%"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.name", "SwapUsage"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.unit", "bytes"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.active", "true"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.name", "NetworkTransmitThroughput"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.tags.4259413593", "source:cloudwatch"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.unit", "bytes"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.name", "DatabaseConnections"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.1.type", "numeric"),
+
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.active", "true"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.name", "DiskQueueDepth"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.tags.4259413593", "source:cloudwatch"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.2.unit", ""),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.name", "FreeStorageSpace"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.3.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.name", "FreeableMemory"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.4.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.name", "MaximumUsedTransactionIDs"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.5.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.name", "NetworkReceiveThroughput"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.6.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.name", "NetworkTransmitThroughput"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.7.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.name", "ReadIOPS"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.8.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.name", "ReadLatency"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.9.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.name", "ReadThroughput"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.10.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.name", "SwapUsage"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.11.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.name", "TransactionLogsDiskUsage"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.12.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.name", "TransactionLogsGeneration"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.13.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.name", "WriteIOPS"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.14.type", "numeric"),
+
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.active", "true"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.name", "WriteLatency"),
+					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.15.type", "numeric"),
+
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.active", "true"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.name", "WriteThroughput"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.tags.1313458811", "app:rds"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.tags.1401442048", "lifecycle:unittest"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.tags.2964981562", "app:postgresql"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.tags.4259413593", "source:cloudwatch"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "metric.16.unit", "bytes"),
 
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "tags.#", "4"),
 					resource.TestCheckResourceAttr("circonus_check.rds_metrics", "tags.2964981562", "app:postgresql"),
@@ -267,117 +181,87 @@ resource "circonus_check" "rds_metrics" {
 
   metric {
     name = "CPUUtilization"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "%%"
   }
 
   metric {
     name = "DatabaseConnections"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "connections"
   }
 
   metric {
     name = "DiskQueueDepth"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
   }
 
   metric {
     name = "FreeStorageSpace"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
   }
 
   metric {
     name = "FreeableMemory"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "MaximumUsedTransactionIDs"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
   }
 
   metric {
     name = "NetworkReceiveThroughput"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "NetworkTransmitThroughput"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "ReadIOPS"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "iops"
   }
 
   metric {
     name = "ReadLatency"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "seconds"
   }
 
   metric {
     name = "ReadThroughput"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "SwapUsage"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "TransactionLogsDiskUsage"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   metric {
     name = "TransactionLogsGeneration"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
   }
 
   metric {
     name = "WriteIOPS"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "iops"
   }
 
   metric {
     name = "WriteLatency"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "seconds"
   }
 
   metric {
     name = "WriteThroughput"
-    tags = "${var.cloudwatch_rds_tags}"
     type = "numeric"
-    unit = "bytes"
   }
 
   tags = "${var.cloudwatch_rds_tags}"