+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
-	"github.com/rileyberton/terraform-provider-circonus/circonus"
+	"github.com/terraform-providers/terraform-provider-circonus/circonus"
 )
 
 func main() {